+++ frontend/src/pages/admin/LoginPage.tsx
@@ -2,7 +2,8 @@ import { useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { api } from '../../api/client'
 import type { Messages } from '../../i18n/types'
-import AdminLayout, { useAdminMessages } from '../../components/admin/AdminLayout'
+import AdminLayout from '../../components/admin/AdminLayout'
+import { useAdminMessages } from '../../components/admin/AdminMessagesContext'
 import AdminLoginForm from '../../components/admin/AdminLoginForm'
 import { getAdminRouteConfig } from '../../admin/paths'
